Fully changed from MakeFile to CMake. Tested with Visual Studio 16 2019, modified source files to not have warnings when building. Fixed qugate_h using 1.0f rather than 1. Fixed qucircuit.test failing because of mismatch with n_children.

+++ src/qugate/qugate.c
@@ -31,7 +31,7 @@ void qugate_z(xiy *mat_2x2) {
 }
 
 void qugate_h(xiy *mat_2x2) {
-    float c = 1.0f / sqrt(2);
+    double c = 1/sqrt(2);
     mat_2x2[0] = xiy_mul_s(xiy_one, c);
     mat_2x2[1] = mat_2x2[0];
     mat_2x2[2] = mat_2x2[0];
